lesson7 finished
