--- src/actions/status.js
@@ -1,5 +1,0 @@
-import {CHECK_ON} from '../constants'
-
-const changeCheckBox = {type: CHECK_ON}
-
-export {changeCheckBox}
