+++ src/store/chats/actions.js
@@ -1,4 +1,4 @@
-import {ADD_NEW_CHAT, DELETE_CHAT} from '../constants'
+import {ADD_NEW_CHAT, DELETE_CHAT} from './constants'
 
 const addNewChat = (id, newChat) => ({
 	type: ADD_NEW_CHAT,
