+++ src/store/chats/constants.js
@@ -1,0 +1,2 @@
+export const ADD_NEW_CHAT = 'ADD_NEW_CHAT'
+export const DELETE_CHAT = 'DELETE_CHAT'
