+++ src/store/chats/reducer.js
@@ -1,22 +1,10 @@
-import {ADD_NEW_CHAT, DELETE_CHAT} from '../constants'
+import {ADD_NEW_CHAT, DELETE_CHAT} from './constants'
 
 const chats = (state = [
-    {
-      name: 'ChatBot',
-      id: 1,
-    },
-    {
-      name: "Vasiliy",
-      id: 2,
-    },
-    {
-      name: "Alisa",
-      id: 3,
-    },
-    {
-      name: "Siri",
-      id: 4,
-    }
+    {name: 'ChatBot',  id: 1},
+    {name: "Vasiliy",  id: 2},
+    {name: "Alisa",    id: 3},
+    {name: "Siri",     id: 4}
   ], action) => {
   switch (action.type) {
     case ADD_NEW_CHAT:
