+++ src/store/messages/actions.js
@@ -1,4 +1,4 @@
-import {ADD_NEW_MESSAGE, ADD_NEW_CHAT_IN_MESSAGES, DELETE_CHAT_IN_MESSAGES} from '../constants'
+import {ADD_NEW_MESSAGE, ADD_NEW_CHAT_IN_MESSAGES, DELETE_CHAT_IN_MESSAGES} from './constants'
 
 export const addNewMessage = (chatId, newMessage) => ({
 	type: ADD_NEW_MESSAGE,
