+++ src/store/messages/constants.js
@@ -1,6 +1,3 @@
-export const CHECK_ON = 'CHECK_ON'
-export const ADD_NEW_CHAT = 'ADD_NEW_CHAT'
-export const DELETE_CHAT = 'DELETE_CHAT'
 export const ADD_NEW_MESSAGE = 'ADD_NEW_MESSAGE'
 export const ADD_NEW_CHAT_IN_MESSAGES = 'ADD_NEW_CHAT_IN_MESSAGES'
 export const DELETE_CHAT_IN_MESSAGES = 'DELETE_CHAT_IN_MESSAGES'
