+++ src/store/messages/reducer.js
@@ -1,4 +1,4 @@
-import {ADD_NEW_MESSAGE,  ADD_NEW_CHAT_IN_MESSAGES, DELETE_CHAT_IN_MESSAGES} from '../constants'
+import {ADD_NEW_MESSAGE,  ADD_NEW_CHAT_IN_MESSAGES, DELETE_CHAT_IN_MESSAGES} from './constants'
 
 const messages = (state = {1:[], 2:[], 3:[], 4:[]}, action) => {
   switch (action.type) {
