+++ src/store/profile/actions.js
@@ -1,0 +1,2 @@
+import {CHECK_ON} from './constants'
+export const changeCheckBox = {type: CHECK_ON}
