+++ src/store/profile/constants.js
@@ -1,0 +1,1 @@
+export const CHECK_ON = 'CHECK_ON'
